Add files via upload
You are given a large integer represented as an integer array of digits, where each digit [i] is the ith digit of the integer. The digits are ordered from most significant to least significant in left-to-right order. The large integer does not contain any leading 0's.

Increment the large integer by one and return the resulting array of digits.